HACK: run generate_derived_key automatically (BROKEN)
AFAIK this library is busted. It's necessary to sometimes create a
"derived" key from the raw bytes passed, but this is done internally
inconsistently, and the method to generate the derived key oneself isn't
exposed/public.

Not sure the the test started failing; I commented it out. Personally
plan to switch to a different library ASAP, even if no longer
maintained.